refactor(postgres): enum handling with createTable
