+++ lib/query-interface.js
@@ -6,6 +6,7 @@ const DataTypes = require('./data-types');
 const SQLiteQueryInterface = require('./dialects/sqlite/query-interface');
 const MSSSQLQueryInterface = require('./dialects/mssql/query-interface');
 const MySQLQueryInterface = require('./dialects/mysql/query-interface');
+const PostgresQueryInterface = require('./dialects/postgres/query-interface');
 const Transaction = require('./transaction');
 const Promise = require('./promise');
 const QueryTypes = require('./query-types');
@@ -155,129 +156,37 @@ class QueryInterface {
    * @return {Promise}
    */
   createTable(tableName, attributes, options, model) {
-    const keys = Object.keys(attributes);
-    const keyLen = keys.length;
     let sql = '';
-    let i = 0;
+    let promise;
 
     options = _.clone(options) || {};
 
-    attributes = _.mapValues(attributes, attribute => {
-      if (!_.isPlainObject(attribute)) {
-        attribute = { type: attribute, allowNull: true };
-      }
+    attributes = _.mapValues(
+      attributes,
+      attribute => this.sequelize.normalizeAttribute(attribute)
+    );
 
-      attribute = this.sequelize.normalizeAttribute(attribute);
-
-      return attribute;
-    });
-
-    // Postgres requires a special SQL command for enums
+    // Postgres requires special SQL commands for ENUM/ENUM[]
     if (this.sequelize.options.dialect === 'postgres') {
-      const promises = [];
-
-      for (i = 0; i < keyLen; i++) {
-        const attribute = attributes[keys[i]];
-        const type = attribute.type;
-
-        if (
-          type instanceof DataTypes.ENUM ||
-          (type instanceof DataTypes.ARRAY && type.type instanceof DataTypes.ENUM) //ARRAY sub type is ENUM
-        ) {
-          sql = this.QueryGenerator.pgListEnums(tableName, attribute.field || keys[i], options);
-          promises.push(this.sequelize.query(
-            sql,
-            _.assign({}, options, { plain: true, raw: true, type: QueryTypes.SELECT })
-          ));
-        }
-      }
-
-      return Promise.all(promises).then(results => {
-        const promises = [];
-        let enumIdx = 0;
-
-        for (i = 0; i < keyLen; i++) {
-          const attribute = attributes[keys[i]];
-          const type = attribute.type;
-          const enumType = type.type || type;
-
-          if (
-            type instanceof DataTypes.ENUM ||
-            (type instanceof DataTypes.ARRAY && enumType instanceof DataTypes.ENUM) //ARRAY sub type is ENUM
-          ) {
-            // If the enum type doesn't exist then create it
-            if (!results[enumIdx]) {
-              sql = this.QueryGenerator.pgEnum(tableName, attribute.field || keys[i], enumType, options);
-              promises.push(this.sequelize.query(
-                sql,
-                _.assign({}, options, { raw: true })
-              ));
-            } else if (!!results[enumIdx] && !!model) {
-              const enumVals = this.QueryGenerator.fromArray(results[enumIdx].enum_value);
-              const vals = enumType.values;
-
-              vals.forEach((value, idx) => {
-                // reset out after/before options since it's for every enum value
-                const valueOptions = _.clone(options);
-                valueOptions.before = null;
-                valueOptions.after = null;
-
-                if (enumVals.indexOf(value) === -1) {
-                  if (vals[idx + 1]) {
-                    valueOptions.before = vals[idx + 1];
-                  }
-                  else if (vals[idx - 1]) {
-                    valueOptions.after = vals[idx - 1];
-                  }
-                  valueOptions.supportsSearchPath = false;
-                  promises.push(this.sequelize.query(this.QueryGenerator.pgEnumAdd(tableName, attribute.field || keys[i], value, valueOptions), valueOptions));
-                }
-              });
-              enumIdx++;
-            }
-          }
-        }
-
-        if (!tableName.schema &&
-          (options.schema || !!model && model._schema)) {
-          tableName = this.QueryGenerator.addSchema({
-            tableName,
-            _schema: !!model && model._schema || options.schema
-          });
-        }
-
-        attributes = this.QueryGenerator.attributesToSQL(attributes, {
-          context: 'createTable'
-        });
-        sql = this.QueryGenerator.createTableQuery(tableName, attributes, options);
-
-        return Promise.all(promises)
-          .tap(() => {
-            // If ENUM processed, then refresh OIDs
-            if (promises.length) {
-              return this.sequelize.dialect.connectionManager._refreshDynamicOIDs();
-            }
-          })
-          .then(() => {
-            return this.sequelize.query(sql, options);
-          });
-      });
+      promise = PostgresQueryInterface.ensureEnums.call(this, tableName, attributes, options, model);
     } else {
-      if (!tableName.schema &&
-        (options.schema || !!model && model._schema)) {
-        tableName = this.QueryGenerator.addSchema({
-          tableName,
-          _schema: !!model && model._schema || options.schema
-        });
-      }
-
-      attributes = this.QueryGenerator.attributesToSQL(attributes, {
-        context: 'createTable'
-      });
-      sql = this.QueryGenerator.createTableQuery(tableName, attributes, options);
-
-      return this.sequelize.query(sql, options);
+      promise = Promise.resolve();
     }
+
+    if (
+      !tableName.schema &&
+      (options.schema || !!model && model._schema)
+    ) {
+      tableName = this.QueryGenerator.addSchema({
+        tableName,
+        _schema: !!model && model._schema || options.schema
+      });
+    }
+
+    attributes = this.QueryGenerator.attributesToSQL(attributes, { context: 'createTable' });
+    sql = this.QueryGenerator.createTableQuery(tableName, attributes, options);
+
+    return promise.then(() => this.sequelize.query(sql, options));
   }
 
   /**
